+++ public/js/gc2/gc2table.js
@@ -361,8 +361,9 @@ var gc2table = (function () {
             }
 
             // Unbind event so no references are left to object
-            object.unbind("clearSelection_" + uid)
-            object.unbind("selected_" + uid)
+            object.unbind("clearSelection_" + uid);
+            object.unbind("selected_" + uid);
+            alert();
         };
 
         assignEventListeners = function () {